% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-chart.R
\name{arc_chart}
\alias{arc_chart}
\title{Start a chart}
\usage{
arc_chart(.data)
}
\arguments{
\item{.data}{A data frame (or similar) the chart's fields will come from.}
}
\value{
An \code{ArcChart}. Pipe it into \code{\link[=set_type]{set_type()}} before \code{\link[=set_x]{set_x()}}/\code{\link[=set_y]{set_y()}}.
}
\description{
Start a chart
}
