% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-map.R
\name{arc_map}
\alias{arc_map}
\title{Start a map}
\usage{
arc_map(basemap = "topo-vector")
}
\arguments{
\item{basemap}{default \code{"topo-vector"}. Defines the basemap the layers draw
over. See \link{Basemaps} for the full set.}
}
\value{
An \link{ArcMap}.
}
\description{
Creates an empty map to pipe \code{\link[=add_layer]{add_layer()}} and the \verb{set_*()} functions
into. Printing it renders the widget.
}
\examples{
arc_map("gray-vector")
}
