% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis-chart-widget.R
\name{arcgisChartOutput}
\alias{arcgisChartOutput}
\alias{renderArcgisChart}
\title{Shiny bindings for arcgis_chart}
\usage{
arcgisChartOutput(outputId, width = "100\%", height = "400px")

renderArcgisChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Defines which output variable the chart is read from.}

\item{width, height}{default \code{"100\%"} and \code{"400px"}. Defines the output size,
passed to \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{Defines the expression that generates the chart.}

\item{env}{default \code{parent.frame()}. Defines where \code{expr} is evaluated.}

\item{quoted}{default \code{FALSE}. Defines whether \code{expr} is already quoted.}
}
\value{
A Shiny output or render function.
}
\description{
Place \code{arcgisChartOutput()} in a Shiny UI and \code{renderArcgisChart()} in the
server.
}
\examples{
arcgisChartOutput("chart")
}
