% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis-map-widget.R
\name{arcgisMapOutput}
\alias{arcgisMapOutput}
\alias{renderArcgisMap}
\title{Shiny bindings for arcgis_map}
\usage{
arcgisMapOutput(outputId, width = "100\%", height = "500px")

renderArcgisMap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Defines which output variable the map is read from.}

\item{width, height}{default \code{"100\%"} and \code{"500px"}. Defines the output size,
passed to \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{Defines the expression that generates the map.}

\item{env}{default \code{parent.frame()}. Defines where \code{expr} is evaluated.}

\item{quoted}{default \code{FALSE}. Defines whether \code{expr} is already quoted.}
}
\value{
A Shiny output or render function.
}
\description{
Place \code{arcgisMapOutput()} in a Shiny UI and \code{renderArcgisMap()} in the
server.
}
\examples{
arcgisMapOutput("map")
}
