% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisviz-package.R
\docType{package}
\name{arcgisviz-package}
\alias{arcgisviz}
\alias{arcgisviz-package}
\title{arcgisviz: Interactive Charts and Maps with the 'ArcGIS Maps SDK'}
\description{
Build interactive charts and maps from data frames and simple feature objects using the 'ArcGIS Maps SDK for JavaScript'. Charts are described with a grammar of graphics style pipeline and render as htmlwidgets, so they work in the viewer, in 'Quarto' documents, and in 'shiny' applications. Data travels to the browser as a client side feature collection, so no feature service is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r.esri.com/arcgisviz/}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

}
\keyword{internal}
