% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-data.R
\name{as_widget}
\alias{as_widget}
\title{Convert a chart to an htmlwidget}
\usage{
as_widget(chart, width = NULL, height = NULL, element_id = NULL)
}
\arguments{
\item{chart}{Defines which chart to render.}

\item{width, height}{default \code{NULL}. Defines the widget size, passed to
\code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}.}

\item{element_id}{default \code{NULL}. Defines the DOM element id to render into.}
}
\value{
An htmlwidget.
}
\description{
Turns an \code{\link[=arc_chart]{arc_chart()}} into a renderable widget. Printing an \code{ArcChart}
calls this for you.
}
\examples{
df <- data.frame(species = c("a", "b", "c"), mass = c(1, 5, 3))

as_widget(arc_col(df, species, mass))
}
