% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{CategoryFormatOptions}
\alias{CategoryFormatOptions}
\title{CategoryFormatOptions}
\usage{
CategoryFormatOptions(type = NA_character_, characterLimit = NA_real_)
}
\arguments{
\item{type}{String.}

\item{characterLimit}{Number.}
}
\value{
An object of class \code{CategoryFormatOptions}.
}
\description{
CategoryFormatOptions
}
\examples{
CategoryFormatOptions(type = "category", characterLimit = 20)
}
