% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-format-options.R
\name{DateTimeFormatOptions}
\alias{DateTimeFormatOptions}
\title{DateTimeFormatOptions}
\usage{
DateTimeFormatOptions(
  type = NA_character_,
  intlOptions = IntlDateTimeFormatOptions(),
  formatPerDateTimeUnit = NULL
)
}
\arguments{
\item{type}{String.}

\item{intlOptions}{A \code{IntlDateTimeFormatOptions} object.}

\item{formatPerDateTimeUnit}{\code{NULL} or a \code{WebChartDateTimeUnitFormatOptions} object.}
}
\value{
An object of class \code{DateTimeFormatOptions}.
}
\description{
How a date axis or label formats its values.
}
\examples{
DateTimeFormatOptions(
  type = "date-time",
  intlOptions = IntlDateTimeFormatOptions(
    year = IntlDateTimeDigitStyle("numeric"),
    month = IntlDateTimeFormatOptionsMonth("short")
  )
)
}
