% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{IFont}
\alias{IFont}
\title{IFont}
\usage{
IFont(
  family = NA_character_,
  size = NA_real_,
  style = IFontStyle(),
  weight = IFontWeight(),
  decoration = IFontDecoration()
)
}
\arguments{
\item{family}{String.}

\item{size}{Number.}

\item{style}{A \code{IFontStyle} enum.}

\item{weight}{A \code{IFontWeight} enum.}

\item{decoration}{A \code{IFontDecoration} enum.}
}
\value{
An object of class \code{IFont}.
}
\description{
IFont
}
\examples{
IFont(
  family = "Avenir Next",
  size = 12,
  style = IFontStyle("normal"),
  weight = IFontWeight("bold")
)
}
