% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IFontWeight}
\alias{IFontWeight}
\title{IFontWeight}
\usage{
IFontWeight(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"bold"}, \code{"bolder"}, \code{"lighter"}, \code{"normal"}, \code{NA}.}
}
\value{
An object of class \code{IFontWeight}.
}
\description{
One of \code{"bold"}, \code{"bolder"}, \code{"lighter"}, \code{"normal"}, \code{NA}.
}
\examples{
IFontWeight("bold")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

