% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{ILegendOptionsOrder}
\alias{ILegendOptionsOrder}
\title{ILegendOptionsOrder}
\usage{
ILegendOptionsOrder(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"ascendingValues"}, \code{"descendingValues"}, \code{NA}.}
}
\value{
An object of class \code{ILegendOptionsOrder}.
}
\description{
One of \code{"ascendingValues"}, \code{"descendingValues"}, \code{NA}.
}
\examples{
ILegendOptionsOrder("ascendingValues")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

