% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeDigitStyle}
\alias{IntlDateTimeDigitStyle}
\title{IntlDateTimeDigitStyle}
\usage{
IntlDateTimeDigitStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"2-digit"}, \code{"numeric"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeDigitStyle}.
}
\description{
One of \code{"2-digit"}, \code{"numeric"}, \code{NA}.
}
\examples{
IntlDateTimeDigitStyle("2-digit")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

