% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeFormatLength}
\alias{IntlDateTimeFormatLength}
\title{IntlDateTimeFormatLength}
\usage{
IntlDateTimeFormatLength(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"full"}, \code{"long"}, \code{"medium"}, \code{"short"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeFormatLength}.
}
\description{
One of \code{"full"}, \code{"long"}, \code{"medium"}, \code{"short"}, \code{NA}.
}
\examples{
IntlDateTimeFormatLength("full")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

