% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeFormatOptionsFormatMatcher}
\alias{IntlDateTimeFormatOptionsFormatMatcher}
\title{IntlDateTimeFormatOptionsFormatMatcher}
\usage{
IntlDateTimeFormatOptionsFormatMatcher(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"basic"}, \code{"best fit"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeFormatOptionsFormatMatcher}.
}
\description{
One of \code{"basic"}, \code{"best fit"}, \code{NA}.
}
\examples{
IntlDateTimeFormatOptionsFormatMatcher("basic")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

