% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeFormatOptionsHourCycle}
\alias{IntlDateTimeFormatOptionsHourCycle}
\title{IntlDateTimeFormatOptionsHourCycle}
\usage{
IntlDateTimeFormatOptionsHourCycle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"h11"}, \code{"h12"}, \code{"h23"}, \code{"h24"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeFormatOptionsHourCycle}.
}
\description{
One of \code{"h11"}, \code{"h12"}, \code{"h23"}, \code{"h24"}, \code{NA}.
}
\examples{
IntlDateTimeFormatOptionsHourCycle("h11")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

