% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlLocaleMatcher}
\alias{IntlLocaleMatcher}
\title{IntlLocaleMatcher}
\usage{
IntlLocaleMatcher(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"best fit"}, \code{"lookup"}, \code{NA}.}
}
\value{
An object of class \code{IntlLocaleMatcher}.
}
\description{
One of \code{"best fit"}, \code{"lookup"}, \code{NA}.
}
\examples{
IntlLocaleMatcher("best fit")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

