% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{IntlNumberFormatOptions}
\alias{IntlNumberFormatOptions}
\title{IntlNumberFormatOptions}
\usage{
IntlNumberFormatOptions(
  localeMatcher = IntlLocaleMatcher(),
  style = IntlNumberFormatOptionsStyle(),
  currency = NA_character_,
  currencyDisplay = IntlNumberFormatOptionsCurrencyDisplay(),
  useGrouping = NA,
  minimumIntegerDigits = NA_real_,
  minimumFractionDigits = NA_real_,
  maximumFractionDigits = NA_real_,
  minimumSignificantDigits = NA_real_,
  maximumSignificantDigits = NA_real_,
  numberingSystem = NA_character_,
  compactDisplay = IntlNumberFormatOptionsCompactDisplay(),
  notation = IntlNumberFormatOptionsNotation(),
  signDisplay = IntlNumberFormatOptionsSignDisplay(),
  unit = NA_character_,
  unitDisplay = IntlFormatWidth(),
  currencySign = IntlNumberFormatOptionsCurrencySign()
)
}
\arguments{
\item{localeMatcher}{A \code{IntlLocaleMatcher} enum.}

\item{style}{A \code{IntlNumberFormatOptionsStyle} enum.}

\item{currency}{String.}

\item{currencyDisplay}{A \code{IntlNumberFormatOptionsCurrencyDisplay} enum.}

\item{useGrouping}{Bool.}

\item{minimumIntegerDigits}{Number.}

\item{minimumFractionDigits}{Number.}

\item{maximumFractionDigits}{Number.}

\item{minimumSignificantDigits}{Number.}

\item{maximumSignificantDigits}{Number.}

\item{numberingSystem}{String.}

\item{compactDisplay}{A \code{IntlNumberFormatOptionsCompactDisplay} enum.}

\item{notation}{A \code{IntlNumberFormatOptionsNotation} enum.}

\item{signDisplay}{A \code{IntlNumberFormatOptionsSignDisplay} enum.}

\item{unit}{String.}

\item{unitDisplay}{A \code{IntlFormatWidth} enum.}

\item{currencySign}{A \code{IntlNumberFormatOptionsCurrencySign} enum.}
}
\value{
An object of class \code{IntlNumberFormatOptions}.
}
\description{
Mirrors JavaScript's own \code{Intl.NumberFormat} options object.
}
\examples{
IntlNumberFormatOptions(
  style = IntlNumberFormatOptionsStyle("decimal"),
  useGrouping = TRUE,
  maximumFractionDigits = 1
)
}
