% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsCompactDisplay}
\alias{IntlNumberFormatOptionsCompactDisplay}
\title{IntlNumberFormatOptionsCompactDisplay}
\usage{
IntlNumberFormatOptionsCompactDisplay(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"long"}, \code{"short"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsCompactDisplay}.
}
\description{
One of \code{"long"}, \code{"short"}, \code{NA}.
}
\examples{
IntlNumberFormatOptionsCompactDisplay("long")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

