% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsCurrencySign}
\alias{IntlNumberFormatOptionsCurrencySign}
\title{IntlNumberFormatOptionsCurrencySign}
\usage{
IntlNumberFormatOptionsCurrencySign(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"accounting"}, \code{"standard"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsCurrencySign}.
}
\description{
One of \code{"accounting"}, \code{"standard"}, \code{NA}.
}
\examples{
IntlNumberFormatOptionsCurrencySign("accounting")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

