% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsNotation}
\alias{IntlNumberFormatOptionsNotation}
\title{IntlNumberFormatOptionsNotation}
\usage{
IntlNumberFormatOptionsNotation(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"compact"}, \code{"engineering"}, \code{"scientific"}, \code{"standard"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsNotation}.
}
\description{
One of \code{"compact"}, \code{"engineering"}, \code{"scientific"}, \code{"standard"}, \code{NA}.
}
\examples{
IntlNumberFormatOptionsNotation("compact")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

