% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsStyle}
\alias{IntlNumberFormatOptionsStyle}
\title{IntlNumberFormatOptionsStyle}
\usage{
IntlNumberFormatOptionsStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"currency"}, \code{"decimal"}, \code{"percent"}, \code{"unit"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsStyle}.
}
\description{
One of \code{"currency"}, \code{"decimal"}, \code{"percent"}, \code{"unit"}, \code{NA}.
}
\examples{
IntlNumberFormatOptionsStyle("currency")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

