% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-statistics.R
\name{IStatisticDefinitionStatisticParameters}
\alias{IStatisticDefinitionStatisticParameters}
\title{IStatisticDefinitionStatisticParameters}
\usage{
IStatisticDefinitionStatisticParameters(
  value = NA_real_,
  orderBy = IStatisticDefinitionStatisticParametersOrderBy()
)
}
\arguments{
\item{value}{Number.}

\item{orderBy}{A \code{IStatisticDefinitionStatisticParametersOrderBy} enum.}
}
\value{
An object of class \code{IStatisticDefinitionStatisticParameters}.
}
\description{
Extra arguments for the statistics that take them - the percentile ones.
}
\examples{
IStatisticDefinitionStatisticParameters(
  value = 0.9,
  orderBy = IStatisticDefinitionStatisticParametersOrderBy("asc")
)
}
