% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IStatisticDefinitionStatisticParametersOrderBy}
\alias{IStatisticDefinitionStatisticParametersOrderBy}
\title{IStatisticDefinitionStatisticParametersOrderBy}
\usage{
IStatisticDefinitionStatisticParametersOrderBy(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"asc"}, \code{"desc"}, \code{NA}.}
}
\value{
An object of class \code{IStatisticDefinitionStatisticParametersOrderBy}.
}
\description{
One of \code{"asc"}, \code{"desc"}, \code{NA}.
}
\examples{
IStatisticDefinitionStatisticParametersOrderBy("asc")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

