% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-format-options.R
\name{NumberFormatOptions}
\alias{NumberFormatOptions}
\title{NumberFormatOptions}
\usage{
NumberFormatOptions(
  type = NA_character_,
  intlOptions = IntlNumberFormatOptions()
)
}
\arguments{
\item{type}{String.}

\item{intlOptions}{A \code{IntlNumberFormatOptions} object.}
}
\value{
An object of class \code{NumberFormatOptions}.
}
\description{
How a numeric axis or label formats its values.
}
\examples{
NumberFormatOptions(
  type = "number",
  intlOptions = IntlNumberFormatOptions(
    useGrouping = TRUE,
    maximumFractionDigits = 0
  )
)
}
