% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{RESTUnits}
\alias{RESTUnits}
\title{RESTUnits}
\usage{
RESTUnits(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"feet"}, \code{"kilometers"}, \code{"meters"}, \code{"miles"}, \code{"nautical-miles"}, \code{"us-nautical-miles"}, \code{NA}.}
}
\value{
An object of class \code{RESTUnits}.
}
\description{
One of \code{"feet"}, \code{"kilometers"}, \code{"meters"}, \code{"miles"}, \code{"nautical-miles"}, \code{"us-nautical-miles"}, \code{NA}.
}
\examples{
RESTUnits("feet")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

