% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-map.R
\name{set_basemap}
\alias{set_basemap}
\title{Set the basemap}
\usage{
set_basemap(map, basemap)
}
\arguments{
\item{map}{Defines which map to modify.}

\item{basemap}{Defines the basemap id. See \link{Basemaps}.}
}
\value{
\code{map}, with the basemap set.
}
\description{
Replaces the imagery the layers draw over.
}
\examples{
set_basemap(arc_map(), "satellite")
}
