% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{SimpleLineSymbolStyle}
\alias{SimpleLineSymbolStyle}
\title{SimpleLineSymbolStyle}
\usage{
SimpleLineSymbolStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"esriSLSDash"}, \code{"esriSLSDashDot"}, \code{"esriSLSDashDotDot"}, \code{"esriSLSDot"}, \code{"esriSLSNull"}, \code{"esriSLSSolid"}, \code{NA}.}
}
\value{
An object of class \code{SimpleLineSymbolStyle}.
}
\description{
One of \code{"esriSLSDash"}, \code{"esriSLSDashDot"}, \code{"esriSLSDashDotDot"}, \code{"esriSLSDot"}, \code{"esriSLSNull"}, \code{"esriSLSSolid"}, \code{NA}.
}
\examples{
SimpleLineSymbolStyle("esriSLSDash")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

