% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{SizePolicyScaleTypes}
\alias{SizePolicyScaleTypes}
\title{SizePolicyScaleTypes}
\usage{
SizePolicyScaleTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"linear"}, \code{"logarithmic"}, \code{NA}.}
}
\value{
An object of class \code{SizePolicyScaleTypes}.
}
\description{
One of \code{"linear"}, \code{"logarithmic"}, \code{NA}.
}
\examples{
SizePolicyScaleTypes("linear")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

