% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{TimeIntervalInfo}
\alias{TimeIntervalInfo}
\title{TimeIntervalInfo}
\usage{
TimeIntervalInfo(unit = WebChartTemporalBinningUnits(), size = NA_real_)
}
\arguments{
\item{unit}{A \code{WebChartTemporalBinningUnits} enum.}

\item{size}{Number.}
}
\value{
An object of class \code{TimeIntervalInfo}.
}
\description{
TimeIntervalInfo
}
\examples{
TimeIntervalInfo(unit = WebChartTemporalBinningUnits("months"), size = 3)
}
