% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-gauge-chart.R
\name{ValueConversion}
\alias{ValueConversion}
\title{ValueConversion}
\value{
An object of class \code{ValueConversion}.
}
\description{
A linear rescaling applied to a gauge's reading before it is drawn.
}
\examples{
# Grams to kilograms.
ValueConversion(factor = 0.001, offset = 0)
}
