% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{WebChartAxisScrollBar}
\alias{WebChartAxisScrollBar}
\title{WebChartAxisScrollBar}
\usage{
WebChartAxisScrollBar(
  visible = NA,
  width = NA_real_,
  color = NULL,
  gripSize = NA_real_,
  margin = NA_real_
)
}
\arguments{
\item{visible}{Bool.}

\item{width}{Number.}

\item{color}{\code{NULL} or a \code{Color} object.}

\item{gripSize}{Number.}

\item{margin}{Number.}
}
\value{
An object of class \code{WebChartAxisScrollBar}.
}
\description{
WebChartAxisScrollBar
}
\examples{
WebChartAxisScrollBar(visible = TRUE, width = 12, gripSize = 6)
}
