% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartCalendarDatePartsUnits}
\alias{WebChartCalendarDatePartsUnits}
\title{WebChartCalendarDatePartsUnits}
\usage{
WebChartCalendarDatePartsUnits(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"dayOfMonth"}, \code{"dayOfWeek"}, \code{"dayOfYear"}, \code{"hourOfDay"}, \code{"monthOfYear"}, \code{"minuteOfDay"}, \code{"weekOfYear"}, \code{"quarterOfYear"}, \code{NA}.}
}
\value{
An object of class \code{WebChartCalendarDatePartsUnits}.
}
\description{
One of \code{"dayOfMonth"}, \code{"dayOfWeek"}, \code{"dayOfYear"}, \code{"hourOfDay"}, \code{"monthOfYear"}, \code{"minuteOfDay"}, \code{"weekOfYear"}, \code{"quarterOfYear"}, \code{NA}.
}
\examples{
WebChartCalendarDatePartsUnits("dayOfMonth")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

