% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartClassBreakTypes}
\alias{WebChartClassBreakTypes}
\title{WebChartClassBreakTypes}
\usage{
WebChartClassBreakTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"equal-interval"}, \code{"quantile"}, \code{"natural-breaks"}, \code{"manual"}, \code{NA}.}
}
\value{
An object of class \code{WebChartClassBreakTypes}.
}
\description{
One of \code{"equal-interval"}, \code{"quantile"}, \code{"natural-breaks"}, \code{"manual"}, \code{NA}.
}
\examples{
WebChartClassBreakTypes("equal-interval")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

