% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartCursorCrosshair}
\alias{WebChartCursorCrosshair}
\title{WebChartCursorCrosshair}
\usage{
WebChartCursorCrosshair(
  type = NA_character_,
  style = NULL,
  verticalLineVisible = NA,
  horizontalLineVisible = NA
)
}
\arguments{
\item{type}{String.}

\item{style}{\code{NULL} or a \code{ISimpleLineSymbol} object.}

\item{verticalLineVisible}{Bool.}

\item{horizontalLineVisible}{Bool.}
}
\value{
An object of class \code{WebChartCursorCrosshair}.
}
\description{
WebChartCursorCrosshair
}
\examples{
WebChartCursorCrosshair(
  type = "chartCursorCrosshair",
  verticalLineVisible = TRUE,
  horizontalLineVisible = FALSE
)
}
