% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartDirectionalDataOrder}
\alias{WebChartDirectionalDataOrder}
\title{WebChartDirectionalDataOrder}
\usage{
WebChartDirectionalDataOrder(
  orderType = WebChartDirectionalDataOrderOrderType(),
  orderBy = WebChartSortOrderKinds(),
  preferLabel = NA
)
}
\arguments{
\item{orderType}{A \code{WebChartDirectionalDataOrderOrderType} enum.}

\item{orderBy}{A \code{WebChartSortOrderKinds} enum.}

\item{preferLabel}{Bool.}
}
\value{
An object of class \code{WebChartDirectionalDataOrder}.
}
\description{
WebChartDirectionalDataOrder
}
\examples{
WebChartDirectionalDataOrder(
  orderType = WebChartDirectionalDataOrderOrderType("arcgis-charts-y-value"),
  orderBy = WebChartSortOrderKinds("DESC")
)
}
