% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartDirectionalDataOrderOrderType}
\alias{WebChartDirectionalDataOrderOrderType}
\title{WebChartDirectionalDataOrderOrderType}
\usage{
WebChartDirectionalDataOrderOrderType(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"arcgis-charts-category"}, \code{"arcgis-charts-mean"}, \code{"arcgis-charts-median"}, \code{"arcgis-charts-y-value"}, \code{NA}.}
}
\value{
An object of class \code{WebChartDirectionalDataOrderOrderType}.
}
\description{
One of \code{"arcgis-charts-category"}, \code{"arcgis-charts-mean"}, \code{"arcgis-charts-median"}, \code{"arcgis-charts-y-value"}, \code{NA}.
}
\examples{
WebChartDirectionalDataOrderOrderType("arcgis-charts-category")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

