% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-gauge-chart.R
\name{WebChartGaugeAxisTick}
\alias{WebChartGaugeAxisTick}
\title{WebChartGaugeAxisTick}
\value{
An object of class \code{WebChartGaugeAxisTick}.
}
\description{
WebChartGaugeAxisTick
}
\examples{
WebChartGaugeAxisTick(type = "chartGaugeAxisTick", visible = TRUE)
}
