% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartGuide}
\alias{WebChartGuide}
\title{WebChartGuide}
\usage{
WebChartGuide(
  type = NA_character_,
  start = NA_character_,
  end = NA_character_,
  style = ISimpleLineSymbol(),
  name = NA_character_,
  label = NULL,
  visible = NA,
  above = NA,
  tooltipReverseColor = NA
)
}
\arguments{
\item{type}{String.}

\item{start}{String or Number.}

\item{end}{String or Number.}

\item{style}{A \code{ISimpleLineSymbol} object or a \code{ISimpleFillSymbol} object.}

\item{name}{String.}

\item{label}{\code{NULL} or a \code{WebChartTextSymbol} object.}

\item{visible}{Bool.}

\item{above}{Bool.}

\item{tooltipReverseColor}{Bool.}
}
\value{
An object of class \code{WebChartGuide}.
}
\description{
A reference line or band drawn across an axis.
}
\examples{
WebChartGuide(
  type = "chartGuide",
  start = 4000,
  name = "Target",
  style = ISimpleLineSymbol(
    style = SimpleLineSymbolStyle("esriSLSDash"),
    color = Color(r = 184, g = 40, b = 40, a = 1),
    width = 1
  )
)
}
