% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartHeatChartHeatRulesTypes}
\alias{WebChartHeatChartHeatRulesTypes}
\title{WebChartHeatChartHeatRulesTypes}
\usage{
WebChartHeatChartHeatRulesTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"gradient"}, \code{"renderer"}, \code{NA}.}
}
\value{
An object of class \code{WebChartHeatChartHeatRulesTypes}.
}
\description{
One of \code{"gradient"}, \code{"renderer"}, \code{NA}.
}
\examples{
WebChartHeatChartHeatRulesTypes("gradient")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

