% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartHeatChartViewTypes}
\alias{WebChartHeatChartViewTypes}
\title{WebChartHeatChartViewTypes}
\usage{
WebChartHeatChartViewTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"SingleCalendarView"}, \code{"SequentialCalendarViews"}, \code{NA}.}
}
\value{
An object of class \code{WebChartHeatChartViewTypes}.
}
\description{
One of \code{"SingleCalendarView"}, \code{"SequentialCalendarViews"}, \code{NA}.
}
\examples{
WebChartHeatChartViewTypes("SingleCalendarView")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

