% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartMultiAxesDataOrder}
\alias{WebChartMultiAxesDataOrder}
\title{WebChartMultiAxesDataOrder}
\usage{
WebChartMultiAxesDataOrder(
  orderType = NA_character_,
  orderByX = NULL,
  orderByY = NULL
)
}
\arguments{
\item{orderType}{String.}

\item{orderByX}{\code{NULL} or String or a \code{WebChartSortOrderKinds} enum.}

\item{orderByY}{\code{NULL} or String or a \code{WebChartSortOrderKinds} enum.}
}
\value{
An object of class \code{WebChartMultiAxesDataOrder}.
}
\description{
WebChartMultiAxesDataOrder
}
\examples{
WebChartMultiAxesDataOrder(
  orderType = "arcgis-charts-multi-axes",
  orderByX = WebChartSortOrderKinds("ASC"),
  orderByY = WebChartSortOrderKinds("DESC")
)
}
