% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-series-shared.R
\name{WebChartNullCategory}
\alias{WebChartNullCategory}
\title{WebChartNullCategory}
\usage{
WebChartNullCategory(text = NA_character_, symbol = NULL)
}
\arguments{
\item{text}{String.}

\item{symbol}{\code{NULL} or a \code{ISimpleFillSymbol} object.}
}
\value{
An object of class \code{WebChartNullCategory}.
}
\description{
How rows with a missing category are labelled and drawn.
}
\examples{
WebChartNullCategory(
  text = "Unknown",
  symbol = ISimpleFillSymbol(color = Color(r = 200, g = 200, b = 200, a = 1))
)
}
