% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{WebChartOrderSeriesBy}
\alias{WebChartOrderSeriesBy}
\title{WebChartOrderSeriesBy}
\usage{
WebChartOrderSeriesBy(preferLabel = NA, orderBy = WebChartSortOrderKinds())
}
\arguments{
\item{preferLabel}{Bool.}

\item{orderBy}{A \code{WebChartSortOrderKinds} enum.}
}
\value{
An object of class \code{WebChartOrderSeriesBy}.
}
\description{
WebChartOrderSeriesBy
}
\examples{
WebChartOrderSeriesBy(
  preferLabel = TRUE,
  orderBy = WebChartSortOrderKinds("DESC")
)
}
