% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-pie-chart.R
\name{WebChartPieChartTick}
\alias{WebChartPieChartTick}
\title{WebChartPieChartTick}
\value{
An object of class \code{WebChartPieChartTick}.
}
\description{
The leader line joining a slice to its label when the label sits outside
the pie.
}
\examples{
WebChartPieChartTick(
  type = "chartPieChartTick",
  visible = TRUE,
  lineSymbol = ISimpleLineSymbol(width = 0.5)
)
}
