% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartQuerySpatialRelationship}
\alias{WebChartQuerySpatialRelationship}
\title{WebChartQuerySpatialRelationship}
\usage{
WebChartQuerySpatialRelationship(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"contains"}, \code{"crosses"}, \code{"disjoint"}, \code{"envelope-intersects"}, \code{"index-intersects"}, \code{"intersects"}, \code{"overlaps"}, \code{"relation"}, \code{"touches"}, \code{"within"}, \code{NA}.}
}
\value{
An object of class \code{WebChartQuerySpatialRelationship}.
}
\description{
One of \code{"contains"}, \code{"crosses"}, \code{"disjoint"}, \code{"envelope-intersects"}, \code{"index-intersects"}, \code{"intersects"}, \code{"overlaps"}, \code{"relation"}, \code{"touches"}, \code{"within"}, \code{NA}.
}
\examples{
WebChartQuerySpatialRelationship("contains")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

