% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-radar-chart.R
\name{WebChartRadarChartAxis}
\alias{WebChartRadarChartAxis}
\title{WebChartRadarChartAxis}
\value{
An object of class \code{WebChartRadarChartAxis}.
}
\description{
A \code{\link[=WebChartAxis]{WebChartAxis()}} with the label orientation only a circular axis needs.
}
\examples{
WebChartRadarChartAxis(
  type = "chartAxis",
  labelsOrientation = WebChartRadarChartAxisLabelsOrientation("circular")
)
}
