% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartRadarChartAxisLabelsOrientation}
\alias{WebChartRadarChartAxisLabelsOrientation}
\title{WebChartRadarChartAxisLabelsOrientation}
\usage{
WebChartRadarChartAxisLabelsOrientation(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"radial"}, \code{"circular"}, \code{"horizontal"}, \code{NA}.}
}
\value{
An object of class \code{WebChartRadarChartAxisLabelsOrientation}.
}
\description{
One of \code{"radial"}, \code{"circular"}, \code{"horizontal"}, \code{NA}.
}
\examples{
WebChartRadarChartAxisLabelsOrientation("radial")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

