% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartSortOrderKinds}
\alias{WebChartSortOrderKinds}
\title{WebChartSortOrderKinds}
\usage{
WebChartSortOrderKinds(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"ASC"}, \code{"DESC"}, \code{NA}.}
}
\value{
An object of class \code{WebChartSortOrderKinds}.
}
\description{
One of \code{"ASC"}, \code{"DESC"}, \code{NA}.
}
\examples{
WebChartSortOrderKinds("ASC")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

