% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-series-shared.R
\name{WebChartTemporalBinning}
\alias{WebChartTemporalBinning}
\title{WebChartTemporalBinning}
\usage{
WebChartTemporalBinning(
  unit = WebChartTemporalBinningUnits(),
  size = NA_real_,
  timeAggregationType = WebChartTimeAggregationTypes(),
  trimIncompleteTimeInterval = NA,
  start = NA_real_,
  end = NA_real_,
  offset = NULL,
  outTimeZone = NA_character_,
  firstDayOfWeek = NA_real_,
  nullPolicy = WebChartNullPolicyTypes()
)
}
\arguments{
\item{unit}{A \code{WebChartTemporalBinningUnits} enum.}

\item{size}{Number.}

\item{timeAggregationType}{A \code{WebChartTimeAggregationTypes} enum.}

\item{trimIncompleteTimeInterval}{Bool.}

\item{start}{Number.}

\item{end}{Number.}

\item{offset}{\code{NULL} or a \code{WebChartTemporalBinningOffset} object.}

\item{outTimeZone}{String.}

\item{firstDayOfWeek}{Number.}

\item{nullPolicy}{A \code{WebChartNullPolicyTypes} enum.}
}
\value{
An object of class \code{WebChartTemporalBinning}.
}
\description{
Groups a date column into intervals before plotting it.
}
\examples{
WebChartTemporalBinning(
  unit = WebChartTemporalBinningUnits("months"),
  size = 1,
  trimIncompleteTimeInterval = TRUE,
  nullPolicy = WebChartNullPolicyTypes("zero")
)
}
