% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-series-shared.R
\name{WebChartTemporalBinningOffset}
\alias{WebChartTemporalBinningOffset}
\title{WebChartTemporalBinningOffset}
\usage{
WebChartTemporalBinningOffset(
  unit = WebChartTemporalBinningUnits(),
  size = NA_real_
)
}
\arguments{
\item{unit}{A \code{WebChartTemporalBinningUnits} enum.}

\item{size}{Number.}
}
\value{
An object of class \code{WebChartTemporalBinningOffset}.
}
\description{
Shifts where temporal bins start - a fiscal year beginning in April rather
than January, say.
}
\examples{
WebChartTemporalBinningOffset(
  unit = WebChartTemporalBinningUnits("months"),
  size = 3
)
}
