% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartTemporalBinningUnits}
\alias{WebChartTemporalBinningUnits}
\title{WebChartTemporalBinningUnits}
\usage{
WebChartTemporalBinningUnits(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"days"}, \code{"hours"}, \code{"minutes"}, \code{"months"}, \code{"quarters"}, \code{"seconds"}, \code{"weeks"}, \code{"years"}, \code{NA}.}
}
\value{
An object of class \code{WebChartTemporalBinningUnits}.
}
\description{
One of \code{"days"}, \code{"hours"}, \code{"minutes"}, \code{"months"}, \code{"quarters"}, \code{"seconds"}, \code{"weeks"}, \code{"years"}, \code{NA}.
}
\examples{
WebChartTemporalBinningUnits("days")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

