% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartTextSymbol}
\alias{WebChartTextSymbol}
\title{WebChartTextSymbol}
\usage{
WebChartTextSymbol(
  type = NA_character_,
  style = NA_character_,
  text = NA_character_,
  color = NULL,
  backgroundColor = NULL,
  borderLineSize = NA_real_,
  borderLineColor = NULL,
  haloSize = NA_real_,
  haloColor = NULL,
  verticalAlignment = WebChartTextSymbolVerticalAlignment(),
  horizontalAlignment = WebChartTextSymbolHorizontalAlignment(),
  rightToLeft = NA,
  kerning = NA,
  font = NULL,
  angle = NA_real_,
  xoffset = NA_character_,
  yoffset = NA_character_
)
}
\arguments{
\item{type}{String.}

\item{style}{String.}

\item{text}{String.}

\item{color}{\code{NULL} or a \code{Color} object.}

\item{backgroundColor}{\code{NULL} or a \code{Color} object.}

\item{borderLineSize}{Number.}

\item{borderLineColor}{\code{NULL} or a \code{Color} object.}

\item{haloSize}{Number.}

\item{haloColor}{\code{NULL} or a \code{Color} object.}

\item{verticalAlignment}{A \code{WebChartTextSymbolVerticalAlignment} enum.}

\item{horizontalAlignment}{A \code{WebChartTextSymbolHorizontalAlignment} enum.}

\item{rightToLeft}{Bool.}

\item{kerning}{Bool.}

\item{font}{\code{NULL} or a \code{IFont} object.}

\item{angle}{Number.}

\item{xoffset}{String or Number.}

\item{yoffset}{String or Number.}
}
\value{
An object of class \code{WebChartTextSymbol}.
}
\description{
WebChartTextSymbol
}
\examples{
WebChartTextSymbol(
  type = "esriTS",
  text = "Mean body mass (g)",
  color = Color(r = 51, g = 51, b = 51, a = 1),
  font = IFont(family = "Avenir Next", size = 12)
)
}
