% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartTimeAggregationTypes}
\alias{WebChartTimeAggregationTypes}
\title{WebChartTimeAggregationTypes}
\usage{
WebChartTimeAggregationTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"equalIntervalsFromEndTime"}, \code{"equalIntervalsFromStartTime"}, \code{NA}.}
}
\value{
An object of class \code{WebChartTimeAggregationTypes}.
}
\description{
One of \code{"equalIntervalsFromEndTime"}, \code{"equalIntervalsFromStartTime"}, \code{NA}.
}
\examples{
WebChartTimeAggregationTypes("equalIntervalsFromEndTime")
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

