% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-chart.R
\name{arc_bar}
\alias{arc_bar}
\title{Bar chart}
\usage{
arc_bar(.data, x, y)
}
\arguments{
\item{.data}{A data frame (or similar) the chart's fields will come from.}

\item{x, y}{Bare column names from \code{.data} (tidy eval).}
}
\value{
An \code{ArcChart}.
}
\description{
Bar chart
}
