% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-chart.R
\name{arc_scatter}
\alias{arc_scatter}
\title{Scatterplot}
\usage{
arc_scatter(.data, x, y)
}
\arguments{
\item{.data}{A data frame (or similar) the chart's fields will come from.}

\item{x, y}{Bare column names from \code{.data} (tidy eval).}
}
\description{
Scatterplot
}
