% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis-chart-widget.R
\name{arcgisChartOutput}
\alias{arcgisChartOutput}
\alias{renderArcgisChart}
\title{Shiny bindings for arcgis_chart}
\usage{
arcgisChartOutput(outputId, width = "100\%", height = "400px")

renderArcgisChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read the chart from.}

\item{width, height}{Sizing, passed to \code{\link[htmlwidgets:shinyWidgetOutput]{htmlwidgets::shinyWidgetOutput()}}.}

\item{expr}{An expression that generates an \code{arcgis_chart}.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This is
useful if you want to save an expression in a variable.}
}
\description{
Shiny bindings for arcgis_chart
}
