% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisviz-package.R
\docType{package}
\name{arcgisviz-package}
\alias{arcgisviz}
\alias{arcgisviz-package}
\title{arcgisviz: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r.esri.com/arcgisviz/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
