% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{CategoryFormatOptions}
\alias{CategoryFormatOptions}
\title{CategoryFormatOptions}
\usage{
CategoryFormatOptions(type = NA_character_, characterLimit = NA_real_)
}
\description{
CategoryFormatOptions
}
