% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-format-options.R
\name{DateTimeFormatOptions}
\alias{DateTimeFormatOptions}
\title{DateTimeFormatOptions}
\usage{
DateTimeFormatOptions(
  type = NA_character_,
  intlOptions = IntlDateTimeFormatOptions(),
  formatPerDateTimeUnit = NULL
)
}
\description{
DateTimeFormatOptions
}
