% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{GaugeChartSubTypes}
\alias{GaugeChartSubTypes}
\title{GaugeChartSubTypes}
\usage{
GaugeChartSubTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"featureGauge"}, \code{"statisticGauge"}, \code{NA}.}
}
\value{
An object of class \code{GaugeChartSubTypes}.
}
\description{
One of \code{"featureGauge"}, \code{"statisticGauge"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

