% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-feature-layer.R
\name{IFeatureLayer}
\alias{IFeatureLayer}
\title{IFeatureLayer}
\description{
A feature layer, either backed by a \code{featureCollection} stored in the web
map or by a service \code{url}. Built by \code{\link[=as_feature_layer]{as_feature_layer()}}.
}
