% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IFeatureLayerFeatureCollectionType}
\alias{IFeatureLayerFeatureCollectionType}
\title{IFeatureLayerFeatureCollectionType}
\usage{
IFeatureLayerFeatureCollectionType(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"markup"}, \code{"notes"}, \code{"route"}, \code{NA}.}
}
\value{
An object of class \code{IFeatureLayerFeatureCollectionType}.
}
\description{
One of \code{"markup"}, \code{"notes"}, \code{"route"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

