% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{IFont}
\alias{IFont}
\title{IFont}
\usage{
IFont(
  family = NA_character_,
  size = NA_real_,
  style = IFontStyle(),
  weight = IFontWeight(),
  decoration = IFontDecoration()
)
}
\description{
IFont
}
