% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IFontDecoration}
\alias{IFontDecoration}
\title{IFontDecoration}
\usage{
IFontDecoration(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"line-through"}, \code{"none"}, \code{"underline"}, \code{NA}.}
}
\value{
An object of class \code{IFontDecoration}.
}
\description{
One of \code{"line-through"}, \code{"none"}, \code{"underline"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

