% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IFontStyle}
\alias{IFontStyle}
\title{IFontStyle}
\usage{
IFontStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"italic"}, \code{"normal"}, \code{"oblique"}, \code{NA}.}
}
\value{
An object of class \code{IFontStyle}.
}
\description{
One of \code{"italic"}, \code{"normal"}, \code{"oblique"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

