% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{IntlDateTimeFormatOptions}
\alias{IntlDateTimeFormatOptions}
\title{IntlDateTimeFormatOptions}
\usage{
IntlDateTimeFormatOptions(
  localeMatcher = IntlLocaleMatcher(),
  weekday = IntlFormatWidth(),
  era = IntlFormatWidth(),
  year = IntlDateTimeDigitStyle(),
  month = IntlDateTimeFormatOptionsMonth(),
  day = IntlDateTimeDigitStyle(),
  hour = IntlDateTimeDigitStyle(),
  minute = IntlDateTimeDigitStyle(),
  second = IntlDateTimeDigitStyle(),
  timeZoneName = IntlDateTimeFormatOptionsTimeZoneName(),
  formatMatcher = IntlDateTimeFormatOptionsFormatMatcher(),
  hour12 = NA,
  timeZone = NA_character_,
  calendar = NA_character_,
  dayPeriod = IntlFormatWidth(),
  numberingSystem = NA_character_,
  dateStyle = IntlDateTimeFormatLength(),
  timeStyle = IntlDateTimeFormatLength(),
  hourCycle = IntlDateTimeFormatOptionsHourCycle(),
  fractionalSecondDigits = NA_integer_
)
}
\description{
IntlDateTimeFormatOptions
}
