% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeFormatOptionsMonth}
\alias{IntlDateTimeFormatOptionsMonth}
\title{IntlDateTimeFormatOptionsMonth}
\usage{
IntlDateTimeFormatOptionsMonth(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"2-digit"}, \code{"long"}, \code{"narrow"}, \code{"numeric"}, \code{"short"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeFormatOptionsMonth}.
}
\description{
One of \code{"2-digit"}, \code{"long"}, \code{"narrow"}, \code{"numeric"}, \code{"short"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

