% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlDateTimeFormatOptionsTimeZoneName}
\alias{IntlDateTimeFormatOptionsTimeZoneName}
\title{IntlDateTimeFormatOptionsTimeZoneName}
\usage{
IntlDateTimeFormatOptionsTimeZoneName(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"long"}, \code{"longGeneric"}, \code{"longOffset"}, \code{"short"}, \code{"shortGeneric"}, \code{"shortOffset"}, \code{NA}.}
}
\value{
An object of class \code{IntlDateTimeFormatOptionsTimeZoneName}.
}
\description{
One of \code{"long"}, \code{"longGeneric"}, \code{"longOffset"}, \code{"short"}, \code{"shortGeneric"}, \code{"shortOffset"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

