% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{IntlNumberFormatOptions}
\alias{IntlNumberFormatOptions}
\title{IntlNumberFormatOptions}
\usage{
IntlNumberFormatOptions(
  localeMatcher = IntlLocaleMatcher(),
  style = IntlNumberFormatOptionsStyle(),
  currency = NA_character_,
  currencyDisplay = IntlNumberFormatOptionsCurrencyDisplay(),
  useGrouping = NA,
  minimumIntegerDigits = NA_real_,
  minimumFractionDigits = NA_real_,
  maximumFractionDigits = NA_real_,
  minimumSignificantDigits = NA_real_,
  maximumSignificantDigits = NA_real_,
  numberingSystem = NA_character_,
  compactDisplay = IntlNumberFormatOptionsCompactDisplay(),
  notation = IntlNumberFormatOptionsNotation(),
  signDisplay = IntlNumberFormatOptionsSignDisplay(),
  unit = NA_character_,
  unitDisplay = IntlFormatWidth(),
  currencySign = IntlNumberFormatOptionsCurrencySign()
)
}
\description{
IntlNumberFormatOptions
}
