% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsCurrencyDisplay}
\alias{IntlNumberFormatOptionsCurrencyDisplay}
\title{IntlNumberFormatOptionsCurrencyDisplay}
\usage{
IntlNumberFormatOptionsCurrencyDisplay(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"code"}, \code{"name"}, \code{"narrowSymbol"}, \code{"symbol"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsCurrencyDisplay}.
}
\description{
One of \code{"code"}, \code{"name"}, \code{"narrowSymbol"}, \code{"symbol"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

