% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-intl-date-time.R
\name{IntlNumberFormatOptionsSignDisplay}
\alias{IntlNumberFormatOptionsSignDisplay}
\title{IntlNumberFormatOptionsSignDisplay}
\usage{
IntlNumberFormatOptionsSignDisplay(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"always"}, \code{"auto"}, \code{"exceptZero"}, \code{"never"}, \code{NA}.}
}
\value{
An object of class \code{IntlNumberFormatOptionsSignDisplay}.
}
\description{
One of \code{"always"}, \code{"auto"}, \code{"exceptZero"}, \code{"never"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

