% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IRendererRotationType}
\alias{IRendererRotationType}
\title{IRendererRotationType}
\usage{
IRendererRotationType(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"arithmetic"}, \code{"geographic"}, \code{NA}.}
}
\value{
An object of class \code{IRendererRotationType}.
}
\description{
One of \code{"arithmetic"}, \code{"geographic"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

