% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{ISimpleFillSymbol}
\alias{ISimpleFillSymbol}
\title{ISimpleFillSymbol}
\usage{
ISimpleFillSymbol(
  type = "esriSFS",
  style = SimpleFillSymbolStyle(),
  color = NULL,
  outline = NULL
)
}
\description{
ISimpleFillSymbol
}
