% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{ISimpleLineSymbol}
\alias{ISimpleLineSymbol}
\title{ISimpleLineSymbol}
\usage{
ISimpleLineSymbol(
  type = NA_character_,
  style = SimpleLineSymbolStyle(),
  color = NULL,
  width = NA_real_
)
}
\description{
ISimpleLineSymbol
}
