% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{ISimpleMarkerSymbol}
\alias{ISimpleMarkerSymbol}
\title{ISimpleMarkerSymbol}
\usage{
ISimpleMarkerSymbol(
  type = "esriSMS",
  style = SimpleMarkerSymbolStyle(),
  color = NULL,
  size = NA_real_,
  outline = NULL,
  angle = NA_real_,
  xoffset = NA_real_,
  yoffset = NA_real_
)
}
\description{
ISimpleMarkerSymbol
}
