% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-statistics.R
\name{IStatisticDefinition}
\alias{IStatisticDefinition}
\title{IStatisticDefinition}
\usage{
IStatisticDefinition(
  statisticType = IStatisticDefinitionStatisticType(),
  statisticParameters = NULL,
  onStatisticField = NA_character_,
  outStatisticFieldName = NA_character_
)
}
\description{
IStatisticDefinition
}
