% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-statistics.R
\name{IStatisticDefinitionStatisticParameters}
\alias{IStatisticDefinitionStatisticParameters}
\title{IStatisticDefinitionStatisticParameters}
\usage{
IStatisticDefinitionStatisticParameters(
  value = NA_real_,
  orderBy = IStatisticDefinitionStatisticParametersOrderBy()
)
}
\description{
IStatisticDefinitionStatisticParameters
}
