% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{IStatisticDefinitionStatisticType}
\alias{IStatisticDefinitionStatisticType}
\title{IStatisticDefinitionStatisticType}
\usage{
IStatisticDefinitionStatisticType(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"avg"}, \code{"centroid-aggregate"}, \code{"convex-hull-aggregate"}, \code{"count"}, \code{"envelope-aggregate"}, \code{"max"}, \code{"min"}, \code{"percentile-continuous"}, \code{"percentile-discrete"}, \code{"stddev"}, \code{"sum"}, \code{"var"}, \code{NA}.}
}
\value{
An object of class \code{IStatisticDefinitionStatisticType}.
}
\description{
One of \code{"avg"}, \code{"centroid-aggregate"}, \code{"convex-hull-aggregate"}, \code{"count"}, \code{"envelope-aggregate"}, \code{"max"}, \code{"min"}, \code{"percentile-continuous"}, \code{"percentile-discrete"}, \code{"stddev"}, \code{"sum"}, \code{"var"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

