% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-format-options.R
\name{NumberFormatOptions}
\alias{NumberFormatOptions}
\title{NumberFormatOptions}
\usage{
NumberFormatOptions(
  type = NA_character_,
  intlOptions = IntlNumberFormatOptions()
)
}
\description{
NumberFormatOptions
}
