% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{SimpleFillSymbolStyle}
\alias{SimpleFillSymbolStyle}
\title{SimpleFillSymbolStyle}
\usage{
SimpleFillSymbolStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"esriSFSBackwardDiagonal"}, \code{"esriSFSCross"}, \code{"esriSFSDiagonalCross"}, \code{"esriSFSForwardDiagonal"}, \code{"esriSFSHorizontal"}, \code{"esriSFSNull"}, \code{"esriSFSSolid"}, \code{"esriSFSVertical"}, \code{NA}.}
}
\value{
An object of class \code{SimpleFillSymbolStyle}.
}
\description{
One of \code{"esriSFSBackwardDiagonal"}, \code{"esriSFSCross"}, \code{"esriSFSDiagonalCross"}, \code{"esriSFSForwardDiagonal"}, \code{"esriSFSHorizontal"}, \code{"esriSFSNull"}, \code{"esriSFSSolid"}, \code{"esriSFSVertical"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

