% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-others.R
\name{SimpleMarkerSymbolStyle}
\alias{SimpleMarkerSymbolStyle}
\title{SimpleMarkerSymbolStyle}
\usage{
SimpleMarkerSymbolStyle(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"esriSMSCircle"}, \code{"esriSMSCross"}, \code{"esriSMSDiamond"}, \code{"esriSMSSquare"}, \code{"esriSMSTriangle"}, \code{"esriSMSX"}, \code{NA}.}
}
\value{
An object of class \code{SimpleMarkerSymbolStyle}.
}
\description{
One of \code{"esriSMSCircle"}, \code{"esriSMSCross"}, \code{"esriSMSDiamond"}, \code{"esriSMSSquare"}, \code{"esriSMSTriangle"}, \code{"esriSMSX"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

