% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-scatter-plot.R
\name{SizePolicy}
\alias{SizePolicy}
\title{SizePolicy}
\usage{
SizePolicy(
  type = NA_character_,
  scaleType = SizePolicyScaleTypes(),
  field = NA_character_,
  minSize = NA_real_,
  maxSize = NA_real_
)
}
\description{
SizePolicy
}
