% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-bar-chart.R
\name{WebChart}
\alias{WebChart}
\title{WebChart}
\usage{
WebChart(
  version = NA_character_,
  type = NA_character_,
  id = NA_character_,
  dataFilters = NULL,
  title = NULL,
  subtitle = NULL,
  footer = NULL,
  background = NULL,
  cursorCrosshair = NULL,
  legend = NULL,
  axes = list(),
  horizontalAxisLabelsBehavior = WebChartLabelBehavior(),
  verticalAxisLabelsBehavior = WebChartLabelBehavior(),
  series = list(),
  rotated = NA,
  stackedType = WebChartStackedKinds(),
  colorMatch = NA,
  chartRenderer = NULL,
  orderOptions = NULL,
  iLayer = NULL
)
}
\description{
WebChart
}
