% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartAxis}
\alias{WebChartAxis}
\title{WebChartAxis}
\usage{
WebChartAxis(
  type = NA_character_,
  visible = NA,
  isLogarithmic = NA,
  title = NULL,
  valueFormat = NumberFormatOptions(),
  minimum = NA_real_,
  maximum = NA_real_,
  grid = NULL,
  guides = list(),
  lineSymbol = NULL,
  labels = WebChartText(),
  scrollbar = NULL,
  displayZeroLine = NA,
  integerOnlyValues = NA,
  displayCursorTooltip = NA,
  buffer = NA,
  tickSpacing = NA_real_,
  dateBaseInterval = NULL
)
}
\description{
WebChartAxis
}
