% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{WebChartAxisScrollBar}
\alias{WebChartAxisScrollBar}
\title{WebChartAxisScrollBar}
\usage{
WebChartAxisScrollBar(
  visible = NA,
  width = NA_real_,
  color = NULL,
  gripSize = NA_real_,
  margin = NA_real_
)
}
\description{
WebChartAxisScrollBar
}
