% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-bar-chart.R
\name{WebChartBarChartSeries}
\alias{WebChartBarChartSeries}
\title{WebChartBarChartSeries}
\usage{
WebChartBarChartSeries(
  type = NA_character_,
  y = NA_character_,
  fillSymbol = NULL,
  hideOversizedStackedLabels = NA,
  hideOversizedSideBySideLabels = NA,
  nullCategory = NULL,
  id = NA_character_,
  visible = NA,
  dataTooltipVisible = NA,
  dataTooltipReverseColor = NA,
  dataTooltipValueFormat = NULL,
  dataTooltipPercentFormat = NULL,
  dataTooltipDateFormat = NULL,
  dataTooltipFontSize = NA_real_,
  name = NA_character_,
  query = NULL,
  x = NA_character_,
  dataLabels = NULL,
  assignToSecondValueAxis = NA,
  binTemporalData = NA,
  temporalBinning = NULL
)
}
\description{
WebChartBarChartSeries
}
