% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartCursorCrosshair}
\alias{WebChartCursorCrosshair}
\title{WebChartCursorCrosshair}
\usage{
WebChartCursorCrosshair(
  type = NA_character_,
  style = NULL,
  verticalLineVisible = NA,
  horizontalLineVisible = NA
)
}
\description{
WebChartCursorCrosshair
}
