% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartDataFilters}
\alias{WebChartDataFilters}
\title{WebChartDataFilters}
\usage{
WebChartDataFilters(
  distance = NA_real_,
  gdbVersion = NA_character_,
  geometry = NULL,
  objectIds = list(),
  spatialRelationship = WebChartQuerySpatialRelationship(),
  timeExtent = list(),
  units = RESTUnits(),
  where = NA_character_
)
}
\description{
WebChartDataFilters
}
