% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartDataTransformations}
\alias{WebChartDataTransformations}
\title{WebChartDataTransformations}
\usage{
WebChartDataTransformations(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"none"}, \code{"logarithmic"}, \code{"squareRoot"}, \code{NA}.}
}
\value{
An object of class \code{WebChartDataTransformations}.
}
\description{
One of \code{"none"}, \code{"logarithmic"}, \code{"squareRoot"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

