% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-format-options.R
\name{WebChartDateTimeUnitFormatOptions}
\alias{WebChartDateTimeUnitFormatOptions}
\title{WebChartDateTimeUnitFormatOptions}
\usage{
WebChartDateTimeUnitFormatOptions(
  year = NULL,
  month = NULL,
  day = NULL,
  hour = NULL,
  minute = NULL,
  second = NULL
)
}
\description{
WebChartDateTimeUnitFormatOptions
}
