% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartDirectionalDataOrder}
\alias{WebChartDirectionalDataOrder}
\title{WebChartDirectionalDataOrder}
\usage{
WebChartDirectionalDataOrder(
  orderType = WebChartDirectionalDataOrderOrderType(),
  orderBy = WebChartSortOrderKinds(),
  preferLabel = NA
)
}
\description{
WebChartDirectionalDataOrder
}
