% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-heat-chart.R
\name{WebChartHeatChartHeatClassBreaksColorRampInfo}
\alias{WebChartHeatChartHeatClassBreaksColorRampInfo}
\title{WebChartHeatChartHeatClassBreaksColorRampInfo}
\description{
WebChartHeatChartHeatClassBreaksColorRampInfo
}
