% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartLabelBehavior}
\alias{WebChartLabelBehavior}
\title{WebChartLabelBehavior}
\usage{
WebChartLabelBehavior(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"hide"}, \code{"rotate"}, \code{"stagger"}, \code{"wrap"}, \code{NA}.}
}
\value{
An object of class \code{WebChartLabelBehavior}.
}
\description{
One of \code{"hide"}, \code{"rotate"}, \code{"stagger"}, \code{"wrap"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

