% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartLegend}
\alias{WebChartLegend}
\title{WebChartLegend}
\usage{
WebChartLegend(
  type = NA_character_,
  visible = NA,
  title = NULL,
  body = NULL,
  position = WebChartLegendPositions(),
  maxHeight = NA_real_,
  roundMarkers = NA
)
}
\description{
WebChartLegend
}
