% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartLegendPositions}
\alias{WebChartLegendPositions}
\title{WebChartLegendPositions}
\usage{
WebChartLegendPositions(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"bottom"}, \code{"left"}, \code{"right"}, \code{"top"}, \code{NA}.}
}
\value{
An object of class \code{WebChartLegendPositions}.
}
\description{
One of \code{"bottom"}, \code{"left"}, \code{"right"}, \code{"top"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

