% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-line-chart.R
\name{WebChartLineChartSeries}
\alias{WebChartLineChartSeries}
\title{WebChartLineChartSeries}
\usage{
WebChartLineChartSeries(
  type = NA_character_,
  y = NA_character_,
  lineSymbol = NULL,
  lineSmoothed = NA,
  showArea = NA,
  markerVisible = NA,
  markerSymbol = NULL,
  areaColor = NULL,
  stackNegativeValuesToBaseline = NA,
  connectLines = NA,
  nullCategory = NULL,
  id = NA_character_,
  visible = NA,
  dataTooltipVisible = NA,
  dataTooltipReverseColor = NA,
  dataTooltipValueFormat = NULL,
  dataTooltipPercentFormat = NULL,
  dataTooltipDateFormat = NULL,
  dataTooltipFontSize = NA_real_,
  name = NA_character_,
  query = NULL,
  x = NA_character_,
  dataLabels = NULL,
  assignToSecondValueAxis = NA,
  binTemporalData = NA,
  temporalBinning = NULL
)
}
\description{
WebChartLineChartSeries
}
