% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartMultiAxesDataOrder}
\alias{WebChartMultiAxesDataOrder}
\title{WebChartMultiAxesDataOrder}
\usage{
WebChartMultiAxesDataOrder(
  orderType = NA_character_,
  orderByX = NULL,
  orderByY = NULL
)
}
\description{
WebChartMultiAxesDataOrder
}
