% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartNullPolicyTypes}
\alias{WebChartNullPolicyTypes}
\title{WebChartNullPolicyTypes}
\usage{
WebChartNullPolicyTypes(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"interpolate"}, \code{"null"}, \code{"zero"}, \code{NA}.}
}
\value{
An object of class \code{WebChartNullPolicyTypes}.
}
\description{
One of \code{"interpolate"}, \code{"null"}, \code{"zero"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

