% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartOrderOptions}
\alias{WebChartOrderOptions}
\title{WebChartOrderOptions}
\usage{
WebChartOrderOptions(series = NULL, data = NULL, orderByFields = character(0))
}
\description{
WebChartOrderOptions
}
