% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{WebChartOrderSeriesBy}
\alias{WebChartOrderSeriesBy}
\title{WebChartOrderSeriesBy}
\usage{
WebChartOrderSeriesBy(preferLabel = NA, orderBy = WebChartSortOrderKinds())
}
\description{
WebChartOrderSeriesBy
}
