% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-scatter-plot.R
\name{WebChartOverlay}
\alias{WebChartOverlay}
\title{WebChartOverlay}
\usage{
WebChartOverlay(
  type = NA_character_,
  created = NA,
  visible = NA,
  symbol = NULL
)
}
\description{
WebChartOverlay
}
