% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-simple.R
\name{WebChartPredefinedLabelsDataOrder}
\alias{WebChartPredefinedLabelsDataOrder}
\title{WebChartPredefinedLabelsDataOrder}
\usage{
WebChartPredefinedLabelsDataOrder(
  orderType = NA_character_,
  orderBy = character(0),
  preferLabel = NA
)
}
\description{
WebChartPredefinedLabelsDataOrder
}
