% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-scatter-plot.R
\name{WebChartScatterplotSeries}
\alias{WebChartScatterplotSeries}
\title{WebChartScatterplotSeries}
\usage{
WebChartScatterplotSeries(
  type = NA_character_,
  y = NA_character_,
  markerSymbol = NULL,
  overlays = NULL,
  sizePolicy = NULL,
  additionalTooltipFields = character(0),
  id = NA_character_,
  visible = NA,
  dataTooltipVisible = NA,
  dataTooltipReverseColor = NA,
  dataTooltipValueFormat = NULL,
  dataTooltipPercentFormat = NULL,
  dataTooltipDateFormat = NULL,
  dataTooltipFontSize = NA_real_,
  name = NA_character_,
  query = NULL,
  x = NA_character_,
  dataLabels = NULL,
  assignToSecondValueAxis = NA
)
}
\description{
WebChartScatterplotSeries
}
