% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-statistics.R
\name{WebChartSeriesQuery}
\alias{WebChartSeriesQuery}
\title{WebChartSeriesQuery}
\usage{
WebChartSeriesQuery(
  outFields = character(0),
  where = NA_character_,
  groupByFieldsForStatistics = character(0),
  outStatistics = list(),
  returnDistinctValues = NA,
  fetchNullValues = NA
)
}
\description{
WebChartSeriesQuery
}
