% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartStackedKinds}
\alias{WebChartStackedKinds}
\title{WebChartStackedKinds}
\usage{
WebChartStackedKinds(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"sideBySide"}, \code{"stacked"}, \code{"stacked100"}, \code{NA}.}
}
\value{
An object of class \code{WebChartStackedKinds}.
}
\description{
One of \code{"sideBySide"}, \code{"stacked"}, \code{"stacked100"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

