% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-series-shared.R
\name{WebChartTemporalBinningOffset}
\alias{WebChartTemporalBinningOffset}
\title{WebChartTemporalBinningOffset}
\usage{
WebChartTemporalBinningOffset(
  unit = WebChartTemporalBinningUnits(),
  size = NA_real_
)
}
\description{
WebChartTemporalBinningOffset
}
