% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-webchart.R
\name{WebChartTextSymbol}
\alias{WebChartTextSymbol}
\title{WebChartTextSymbol}
\usage{
WebChartTextSymbol(
  type = NA_character_,
  style = NA_character_,
  text = NA_character_,
  color = NULL,
  backgroundColor = NULL,
  borderLineSize = NA_real_,
  borderLineColor = NULL,
  haloSize = NA_real_,
  haloColor = NULL,
  verticalAlignment = WebChartTextSymbolVerticalAlignment(),
  horizontalAlignment = WebChartTextSymbolHorizontalAlignment(),
  rightToLeft = NA,
  kerning = NA,
  font = NULL,
  angle = NA_real_,
  xoffset = NA_character_,
  yoffset = NA_character_
)
}
\description{
WebChartTextSymbol
}
