% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartTextSymbolHorizontalAlignment}
\alias{WebChartTextSymbolHorizontalAlignment}
\title{WebChartTextSymbolHorizontalAlignment}
\usage{
WebChartTextSymbolHorizontalAlignment(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"center"}, \code{"justify"}, \code{"left"}, \code{"right"}, \code{NA}.}
}
\value{
An object of class \code{WebChartTextSymbolHorizontalAlignment}.
}
\description{
One of \code{"center"}, \code{"justify"}, \code{"left"}, \code{"right"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

