% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums-web-chart.R
\name{WebChartTextSymbolVerticalAlignment}
\alias{WebChartTextSymbolVerticalAlignment}
\title{WebChartTextSymbolVerticalAlignment}
\usage{
WebChartTextSymbolVerticalAlignment(value = NA_character_)
}
\arguments{
\item{value}{String. One of \code{"baseline"}, \code{"bottom"}, \code{"middle"}, \code{"top"}, \code{NA}.}
}
\value{
An object of class \code{WebChartTextSymbolVerticalAlignment}.
}
\description{
One of \code{"baseline"}, \code{"bottom"}, \code{"middle"}, \code{"top"}, \code{NA}.
}
\section{Additional properties}{

\describe{
\item{\code{@variants}}{Character vector. The values this enum allows.}

\item{\code{@allow_na}}{Bool. Whether \code{NA_character_} is allowed.}
}}

